Change custom data type to float

+++ eq-author-api/schema/typeDefs.js
@@ -429,7 +429,7 @@ type MinValueValidationRule implements ValidationRule {
   id: ID!
   enabled: Boolean!
   inclusive: Boolean!
-  custom: Int
+  custom: Float
   previousAnswer: BasicAnswer
   entityType: ValidationRuleEntityType!
   validationErrorInfo: ValidationErrorInfo
@@ -439,7 +439,7 @@ type MaxValueValidationRule implements ValidationRule {
   id: ID!
   enabled: Boolean!
   inclusive: Boolean!
-  custom: Int
+  custom: Float
   previousAnswer: BasicAnswer
   entityType: ValidationRuleEntityType!
   validationErrorInfo: ValidationErrorInfo
@@ -503,7 +503,7 @@ type TotalValidationRule implements ValidationRule {
   id: ID!
   enabled: Boolean!
   entityType: ValidationRuleEntityType!
-  custom: Int
+  custom: Float
   previousAnswer: Answer
   condition: ValidationCondition!
   validationErrorInfo: ValidationErrorInfo
@@ -1453,14 +1453,14 @@ input UpdateValidationRuleInput {
 
 input UpdateMinValueInput {
   inclusive: Boolean!
-  custom: Int
+  custom: Float
   entityType: ValidationRuleEntityType
   previousAnswer: ID
 }
 
 input UpdateMaxValueInput {
   inclusive: Boolean!
-  custom: Int
+  custom: Float
   entityType: ValidationRuleEntityType
   previousAnswer: ID
 }
@@ -1498,7 +1498,7 @@ input DurationInput {
 
 input UpdateTotalValidationInput {
   entityType: ValidationRuleEntityType!
-  custom: Int
+  custom: Float
   previousAnswer: ID
   condition: ValidationCondition!
   allowUnanswered: Boolean
